update nwb_table_readme.py script with options for downloading files, file size limit, number of dandisets being processed and readme file being updated or not after summary file is created

+++ nwb_table_readme.py
@@ -12,11 +12,12 @@ from nwbinspector.register_checks import Importance
 from nwbinspector.inspector_tools import save_report, format_messages, MessageFormatter
 from dandi import download
 
-def create_dandiset_summary(args_sizelimit=None):
-    if args_sizelimit:
-        hard_limit = 1000000000
-    else:
+def create_dandiset_summary(args_nodownload=None,args_nosizelimit=None,args_dandisetlimit=None,args_updatereadme=None):
+    # if users want no limit on file size for downloading
+    if args_nosizelimit:
         hard_limit = 100000000000
+    else:
+        hard_limit = 1000000000
     json_file = '.dandi/assets.json'
     # directory for dandisets
     root_folder = '/tmp/dandisets'
@@ -30,12 +31,15 @@ def create_dandiset_summary(args_sizelimit=None):
         dl.install(source='https://github.com/dandi/dandisets.git', path=root_folder, recursive=True)
 
     dandiset_folder_name = sorted([item for item in os.listdir(root_folder) if item.startswith('0')])
+
+    # if user only wants to run the script for 10 dandisets
+    if args_dandisetlimit:
+        dandiset_folder_name = dandiset_folder_name[:10]
     yaml_file = 'dandiset.yaml'
 
     yaml_df_flatten = ['identifier','citation','name','assetsSummary.numberOfBytes','assetsSummary.numberOfFiles',
                        'assetsSummary.numberOfSubjects','assetsSummary.variableMeasured','keywords','schemaKey','schemaVersion','url','version']
     tmp_col = ['species','data_type','doi_link','nwb_version','validation_summary','max_file_size','min_file_size']
-    readme_table = ['identifier','data_type','num_files','num_bytes','dandiset_schemaver','url', 'nwb_version','validation_summary','max_file_size','min_file_size']
 
     dandi_metadata = pd.DataFrame()
     nanval = math.nan
@@ -65,6 +69,7 @@ def create_dandiset_summary(args_sizelimit=None):
         largest_file_size = nanval
         smallest_size_lst = [nanval,nanval]
         validation_summary = nanval
+
         if data_type != nanval and 'NWB' in str(data_type):
             # get the json file that has individual files info
             with open(os.path.join(root_folder, dandiset_name, json_file)) as f:
@@ -103,32 +108,37 @@ def create_dandiset_summary(args_sizelimit=None):
                     break
             largest_file_size = json_df['size'].iloc[-1]
             report_message = []
-            # in case files larger than the hard_limit
-            validation_summary = 'NULL_FILE_LIMIT'
-            # only download files that has size lower than the hard limit
-            for i in range(len(path_lst)):
-                if smallest_size_lst[i] < hard_limit:
-                    # download files
-                    nwb_path = download_nwb_with_path(url_lst[i],path_lst[i])
-                    # get nwb_version
-                    nwb_version = get_nwb_version(nwb_path)
-                    # validate file here first
-                    try:
-                        report_message.extend(list(inspect_nwb(nwbfile_path=nwb_path,
-                                                               importance_threshold=Importance.BEST_PRACTICE_VIOLATION)))
+            # if user doesn't want to download files
+            if args_nodownload:
+                validation_summary = 'NOT_DOWNLOADED'
+            # only download files whose sizes are lower than the hard limit
+            else:
+                # in case files larger than the hard_limit and not downloaded
+                validation_summary = 'NULL_FILE_LIMIT'
+                for i in range(len(path_lst)):
+                    if smallest_size_lst[i] < hard_limit:
+                        # download files
+                        nwb_path = download_nwb_with_path(url_lst[i],path_lst[i])
+                        # get nwb_version
+                        nwb_version = get_nwb_version(nwb_path)
+                        # validate file here first
+                        try:
+                            report_message.extend(list(inspect_nwb(nwbfile_path=nwb_path,
+                                                                   importance_threshold=Importance.BEST_PRACTICE_VIOLATION)))
 
-                        validation_summary = nwb_inspector_message_format(report_message, dandiset_name)
-                    except ValueError:
-                        validation_summary = 'UNABLE'
-                        pass
-                    # uninstall file
-                    os.unlink(nwb_path)
+                            validation_summary = nwb_inspector_message_format(report_message, dandiset_name)
+                        except ValueError:
+                            validation_summary = 'UNABLE'
+                            pass
+                        # uninstall file
+                        os.unlink(nwb_path)
         # concatenate the additional variables to the flattened pdf
         yaml_df = pd.concat([yaml_df, pd.DataFrame([[species_name,data_type,doi_link,nwb_version,validation_summary,largest_file_size,smallest_size_lst[0]]],
                                                    index=yaml_df.index,columns=tmp_col)],axis=1)
 
         # concatenate every newly read dandiset metadata dataframe
         dandi_metadata = pd.concat([dandi_metadata,yaml_df],axis=0,ignore_index=True)
+        # in case script crashes
         dandi_metadata.to_csv('dandiset_summary_tmp.csv')
         f.close()
 
@@ -137,14 +147,12 @@ def create_dandiset_summary(args_sizelimit=None):
     dandi_metadata_final = dandi_metadata[yaml_df_flatten].sort_values(by=['identifier'],ignore_index=True)
     dandi_metadata_final.rename(columns={'assetsSummary.numberOfBytes':'num_bytes','assetsSummary.numberOfFiles':'num_files','assetsSummary.numberOfSubjects':'numb_subjects',
                                          'assetsSummary.variableMeasured':'variableMeasured', 'schemaVersion':'dandiset_schemaver'},inplace=True)
-    # get table for readme
-    dandi_metadata_readme = dandi_metadata_final[readme_table]
     # save table to csv
     dandi_metadata_final.to_csv('dandiset_summary.csv')
-    dandi_metadata_readme.to_csv('dandiset_summary_readme.csv')
 
-    # # remove the cloned dandisets folder
+    # remove the cloned dandisets folder
     dl.remove(dataset=root_folder)
+    return args_updatereadme
 
 def download_nwb_with_path(dandi_url,nwb_file_name):
     # the argument nwb_version is in case the function exits with error
@@ -243,7 +251,7 @@ def update_readme():
             readme += '- NWB version: ' + nwb_pd['nwb_version'].iloc[row] + '\n\n'
         if not pd.isna(nwb_pd['num_bytes'].iloc[row]):
             readme += '- Dandiset size (in bytes): ' + str(nwb_pd['num_bytes'].iloc[row]) + '\n\n'
-        if not nwb_pd['validation_summary'].iloc[row] in ['NULL_FILE_LIMIT', 'UNABLE']:
+        if not nwb_pd['validation_summary'].iloc[row] in ['NULL_FILE_LIMIT', 'UNABLE', 'NOT_DOWNLOADED']:
             readme += '- Validation results summary: [' + nwb_pd['validation_summary'].iloc[
                 row] + ']' + '(%s.txt) \n\n' % (validation_file)
         else:
@@ -263,11 +271,20 @@ def update_readme():
     rmd.close()
 
 if __name__ == '__main__':
-    # option for filesize limit
+    # options for users
     parser = argparse.ArgumentParser(description='cap limit on downloaded file size')
-    parser.add_argument('--size_limit', default=False, action='store_true', help='only download files whose sizes are less than a defined hard limit')
+    parser.add_argument('--no_download', default=False, action='store_true',
+                        help='files will not be downloaded for testing if so chosen')
+    parser.add_argument('--no_sizelimit', default=False, action='store_true',
+                        help='no size limit will be capped for downloading files if so chosen')
+    parser.add_argument('--dandiset_limit', default=False, action='store_true',
+                        help='only process first 10 dandisets if so chosen')
+    parser.add_argument('--update_readme_option', default=False, action='store_true',
+                        help='update readme file after summary file is created')
     args = parser.parse_args()
-    create_dandiset_summary(args.size_limit)
-    update_readme()
+    update_readme_option=create_dandiset_summary(args.no_download,args.no_sizelimit,args.dandiset_limit,args.update_readme_option)
+
+    if update_readme_option:
+        update_readme()
